+++ discord/channel.py
@@ -30,7 +30,7 @@ from typing import Callable, Dict, List, Optional, TYPE_CHECKING, Union, overloa
 
 import discord.abc
 from .permissions import PermissionOverwrite, Permissions
-from .enums import ChannelType, PrivacyLevel, try_enum, VoiceRegion, VideoQualityMode
+from .enums import ChannelType, StagePrivacyLevel, try_enum, VoiceRegion, VideoQualityMode
 from .mixins import Hashable
 from . import utils
 from .asset import Asset
@@ -922,7 +922,7 @@ class StageChannel(VocalGuildChannel):
         """
         return utils.get(self.guild.stage_instances, channel_id=self.id)
 
-    async def create_instance(self, *, topic: str, privacy_level: PrivacyLevel = utils.MISSING) -> StageInstance:
+    async def create_instance(self, *, topic: str, privacy_level: StagePrivacyLevel = utils.MISSING) -> StageInstance:
         """|coro|
 
         Create a stage instance.
@@ -957,7 +957,7 @@ class StageChannel(VocalGuildChannel):
         }
 
         if privacy_level is not utils.MISSING:
-            if not isinstance(privacy_level, PrivacyLevel):
+            if not isinstance(privacy_level, StagePrivacyLevel):
                 raise InvalidArgument('privacy_level field must be of type PrivacyLevel')
 
             payload['privacy_level'] = privacy_level.value
