+++ discord/enums.py
@@ -50,6 +50,7 @@ __all__ = (
     'VideoQualityMode',
     'ComponentType',
     'ButtonStyle',
+    'StagePrivacyLevel',
 )
 
 def _create_value_cls(name):
@@ -479,7 +480,8 @@ class ButtonStyle(Enum):
 
     def __int__(self):
         return self.value
-class PrivacyLevel(Enum):
+
+class StagePrivacyLevel(Enum):
     public = 1
     closed = 2
     guild_only = 2
